refactor(web): rename voice-chat test seeder and assertion files
